Bug #29486538 MYSQLD CRASHES IN CALC_NDB_BLOB_LEN WHILE TRUNCATING TINYBLOBS
Reviewed-by: Frazer Clement <frazer.clement@oracle.com>

TINYTEXT values are truncated in ndb_restore by truncating to the
nearest well-formed character to the truncation point. This patch
adds support for the truncation of TINYBLOB fields which do not have
a charset. For TINYBLOBS, the my_charset_bin BINARY charset is used
to calculate the character length.

The test cases which were disabled pending this fix are now re-
enabled.