Bug #28966733 NDB_FULL_INDEX_MEMORY TEST FAIL
The embedded libmysqld is not reconnecting promptly when data nodes
are restarted, causing some failures in this testcase.

This appears to be due to problems with deadlocks in the implementation
of getaddrinfo.

As the embedded libmysqld is not supported in the latest versions, this
is not fully debugged, and instead this test is modified to be skipped
on embedded runs.

Reviewed-by: Martin Skold <martin.skold@oracle.com>