BUG#29140212 HANDLE_FATAL_SIGNAL (SIG=11) IN SAFE_MUTEX_LOCK | MYSYS/THR_MUTEX.CC:70
When the server is started with performance schema parameters that causes
an out of memory condition, it fails to start.

The root cause is in initialize_performance_schema() :
Out of memory conditions are supposed to be handled gracefully,
but the initialization in this case is incomplete,
as init_pfs_plugin_table() is not called.

This in turns causes the mutex LOCK_pfs_share_list to be left
in an uninitialized state, causing failures upon usage.

The fix is as follows:

First, improve the SAFE_MUTEX (DEBUG) build to catch cleanly,
with added asserts, uses of uninitialized mutexes.
This helps to diagnose similar errors, pointing closer to the root cause.

Second, rework initialize_performance_schema() to always
call init_pfs_plugin_table(), to initialize cleanly
the performance schema plugin table sub system,
even when some instrumentation is not installed due to
out of memory conditions.

With this fix, the server gracefully handles out of memory conditions
that occur during the performance schema initialization,
leaving the server in a functional state.

Approved by: Chris Powers <chris.powers@oracle.com>