Bug #29635702 : WL#12680: NODES CRASHED DURING SHUTDOWN WITH ERROR "QMGR (LINE: 3398) 0X00000006
Cause:
------

If we miss more than 10 heartbeats from another node, we crash since
we are in the process of stopping and the STOP has mostly likely hanged.
However, 10 heartbeats could mean any time interval depending on how
HeartbeatIntervalDbDb is configured. Hence, we can crash much earlier or
later than required.

Fix:
----

Crash if we've missed heartbeats for 2 minutes and there are at least
10 heartbeat failures.
STOP usually hangs due to GCP STOP errors. And, it takes about 2 minutes
to crash on GCP STOP. Hence, we set the time interval to 2 minutes so that
we can crash on GCP STOP if that's the cause of error.

Reviewed-by: Mauritz Sundell <mauritz.sundell@oracle.com>