BUG#29648393 - X PLUGIN TESTS ARE NOT CORRECTLY CHECKING FOR ITS PRESENCE
Description
===========

Building MySQL Server with option -DWITH_MYSQLX=0, excludes X Plugin from
the build, still MTR tests are not ready to run without it. The condition
that checks if X Plugin is there (in mysql-test/include/have_mysqlx_plugin.inc)
is invalid.

Fix
===

"have_mysqlx_plugin.inc" was updated, to handle this case correctly.
Additionally the "mysqlxclient" library is going to be always built.
This is required because router added a dependence to the libmysqlxclient,
thus mysqlx plugin can't disable the library.

RB: 22031
Reviewed by: Grzegorz Szwarc <grzegorz.szwarc@oracle.com>
Reviewed by: Georgi Kodinov <georgi.kodinov@oracle.com>