Bug#29719713 LOCK_ORDER SYSTEM VARIABLES INVISIBLE TO SHOW VARIABLES
When building WITH_LOCK_ORDER,
the output of SHOW VARIABLES did not show the following variables:
- lock-order-dependencies
- lock-order-extra-dependencies
- lock-order-output-directory

The root cause is the use of flag NOT_VISIBLE when defining
the variables, which is not justified.

The fix is to remove the NOT_VISIBLE flag in
the system variables definition.

Approved by: Chris Powers <chris.powers@oracle.com>