WL#10164 Add ndb_schema_dist_lock_wait_timeout
Add new variable ndb-schema-dist-lock-wait-timeout which controls how
long the ndbcluster binlog injector thread will wait for locks to be
acquired while handling schema operations.

Acquiring these locks are necessary to update the data dictionary (DD)
with latest metadata. While it's important to update the DD, it's even
more important that the ndbcluster binlog injector thread does not block
for an extended amount of time but instead rather skip updating the DD
if lock acquisition fails. By using this variable, it's possible to
configure how long it's acceptable for the thread to wait.

Change-Id: Ia1480099e4f7eeace52cc92a5fa9e47927c2fd34