Bug#29744353 TEST CASE NDB_RESTORE_MISC AND NDB_RESTORE_COMPAT_COMPRESSION FAILS IN PB2
ndb_restore uses exit() to exit the program in a few places. This is not a good
practice when there is dynamic memory allocation in the code, because exit()
will terminate the process immediately without freeing the memory and leads
to the memory leak.

Fix by replacing exit() with exitHandler()

Change-Id: Ie5107276360db74aa196c28e976ba7150c9ce088