Bug 29746812 - MYSQLROUTER SOMETIMES DEADLOCK ON SHUTDOWN.
There was a race condition in the log_reopen_thread() that uses condition variable.
It could not get notified about the application shutdown and deadlock on cond.wait()

This patch makes sure the proper locking is done while notifying the condition variable
so that it always see the proper state of the predicate in the cond.wait(mutex, predicate).

RB: 22042