Bug #29713867 TESTCASE NDB_RPL_GCOL_RECORD_LAYOUT FAILS IN VALGRIND
Virtual blob columns are not included in Binlog events, but their
length is used as part of calculating the amount of memory needed
to store a row event.

So the Binlog Injector is changed to set the Blob length to zero
for virtual Blob columns.

Additionally, the testcase is extended to cover a situation with
2 virtual blob columns in one table.

Approved by : John David Duncan <john.duncan@oracle.com>