Bug #29440208 'INNODB.DDL_CRASH_ALTER_TABLE_PARTITION' TEST FAILS ON PB2 BRANCH
Problem:
While doing the crash recovery, the log pages that are not found in the buffer pool hash
are added to the read ahead buffer area array. These pages are then subsequently read
into the buffer pool during recovery.

The problem happens if there are no pages that need to be read into the buffer pool.
However, before calling the routine to read the pages, no check is made to ensure that
there is atleast one page that needs to be read into the buffer pool. This leads to the
situation where we end up reading an incorrect position in an uninitialized array.

Solution:
Check if there are any pages to be read during recovery before issuing a read request

RB# 22058
Reviewed By: Kevin Lewis (kevin.lewis@oracle.com)