Bug#29224561: USE "SORTED_RESULT" OF MYSQL-TEST , CAUSE ERROR RESULT               ( FOR TYPE OF BINARY )
Issue:
------
When --sorted_result is used for a select on a table which has
column(s) of binary type, the result which is printed is
truncated.

Fix:
----
Mysqltest traverses the result set till characters exist, and
inserts the rows line by line into an array before sorting them.
With binary data, there might be null characters (0x00) as padding
which mysqltest ignores, and all the data after that is truncated.
Instead of traversing the result set till characters exist, now
mysqltest traverses from the start to end of the result set using
the size of it. Also, std::string is used instead of c strings
to ensure that null characters which are part of the string in
char type are also included.

Change-Id: If8b68298f07194812042c0b52ba795279067831d