Bug #29466089 RETRY FAILURE OF UPDATE NDB_APPLY STATUS DUE TO REDO OVERLOAD
Reviewed-by: Frazer Clement <frazer.clement@oracle.com>

Epoch restore in ndb_restore fails because of a temporary redo error.
The epoch restore is not retried, so this results in ndb_restore
erroring out.

This is fixed by adding retries to the epoch update code in ndb_restore.
A previously deleted error insert is reimplemented in DBLQH to simulate a
temporary out-of-redo error. A test case is added in autotest for
epoch restore with temporary errors.