BUG#29759391 MYSQLROUTER SOMETIMES CRASHES ON SHUTDOWN WHEN COULD NOT SPAWN A NEW THREAD.
There is a test in the router suite that checks how the Router handles the
scenario when it can't spawn another thread.
This test fails sometimes on pb2 as the Router binary crashes receiving
SIGABORT.

When creating a thread in the same step we emplace it in the
vector. That results i half-initailized or uninitialized object
being added to the vector which then crashes when the vector is being destroyed.

This patch splits the thread creation and adding it to the vector into 2 steps.

RB: 22076