Bug #29754967 _GET_COMMUNICATION_PROTOCOL() FAILS IF ONE OR MORE MEMBERS ARE RECOVERING
Problem and analysis
========================================================================
The UDF group_replication_get_communication_protocol(), used to inspect
the group's communication protocol version at runtime, fails if one or
more group members are not ONLINE.

There is no particular technical reason for this behaviour, so we can
relax the original implementation behaviour.

Solution
========================================================================
Execute the UDF as long as the member is ONLINE and in contact with a
majority.

Reviewed-by: Pedro Gomes <pedro.gomes@oracle.com>
Reviewed-by: Tiago Jorge <tiago.jorge@oracle.com>
RB: 22082