Bug #29004362	LOCK->TRX->DUPLICATES CAN NOT BE USED TO DETERMINE IF LOCK IS CREATED BY IODKU
When purging a deleted row, we determine if an old lock of on the record should be inherited as a gap lock, or can be safely ignored.
This is important mainly for constraint checks in which we lock rows to avoid conflicts,
and even in READ COMMITTED such locks are needed until we finish the insert statement processing.
The type of the lock (S or X) requested during constraint check depends on variant of INSERT statement (IODKU, and REPLACE need X).
There was a heuristic, which wrongly looked at the type of the query being currently performed by a transaction
to determine which type of locks needs to be inherited.
This was wrong, because requests for a weak lock (S) can be satisfied by already granted stronger locks (X),
thus we need to inherit also such stronger locks.
This patch fixes the heuristic, so that it inherits all locks if the transaction is in the middle of a statement requiring constraint check.

It also adds MTRs which demonstrate:
- how inheriting only the locks of the type required for the current query could lead to duplication in a unique index.
- that it is not enough to only inherit LOCK_X
- that it is not enough to only inherit LOCK_S
- what locks we take when doing {INSERT,REPLACE} x {no-conflict,conflict,"conflict" with delete-marked row}^{PRIMARY,UNIQUE KEY}

Reviewed-by: Debarun Banerjee <debarun.banerjee@oracle.com>
RB:21102