Bug #29594082: INCORRECT AND DANGEROUS ERROR MESSAGE FROM MY-013236 RB#22109
The server was outputting the same message regardless if the directory
was created by the server at initialize time or not.
Added a flag if the data directory is created by the server or not.
Added a new error message if it isn't. Reworded slightly the original
error message.
Test added.