Bug #29752685 ADD MANUAL PAGES TO MYSQL ROUTER
The commit adds the current state of the man pages. It also adds them
to the proper sections in rpm/deb packaging specs. These two things have
to be done in sync.

Similar to NDB man pages, they are removed from the build if building
WITH_ROUTER=0. In order for this to work, had to move cmake processing
of the 'man' subdirectory to after 'router' where WITH_ROUTER is set.

Change-Id: I95373702e49706eb6ac47db50c6e878df6238a5e