Bug#29699347 ITEM_FLOAT C'TOR DOES NOT SET MAX_LENGTH MEMBER
One of the Item_float constructors did not set the max_length
member variable as required by the semantics of this variable.
(Should only be zero for a zero length string, and possibly a
NULL valued Item)

This patch fixes that by setting max_length to the value provided
by the function float_length(). This is similar to what the
existing methods set_data_type_double() and set_data_type_float()
does.

Reviewed by: Jon Olav Hauglid <jon.hauglid@oracle.com>