Bug#29382585 getpeername() errno 107 at startup
Removed the call to getpeername() as the client address is already
known from accept().

Background
==========

At startup getpeername() was called to get the client peer address
for logging purposes.

In case the client socket is already shutdown() getpeername()
would return ENOTCONNECTED which led to a uncaught exception which
lead to a shutdown a plugin start.

Extra cleanup
=============

change 'void *' in socketop.inetntop() to 'const void *'

RB: 22136