Bug #25289359 DML/DDL ON LARGE FULLTEXT TABLES CAUSE SEMAPHORE TIMEOUTS AND ASSERTION/SUICIDE
BACKGROUND

While data from fts cache is being synced in the background, we allow
fts cache lock to be released while internal sql that handles insert to
aux table is performed.

PROBLEM

When the data from the cache that needs to be synced becomes larger than
the value of fts_max_cache_size(configurable) we don't allow cache lock
to be released anymore by the sync as it would never finish otherwise.

FIX

A middlepath would be to basically
 1) Extend the timeout and let sync go on with cache lock held avoiding
    monitor thread to crash after 600 seconds
 2) If the sync takes longer then we allow cache lock to be released by
    sync.

Reviewed by: bin su<bin.x.su@oracle.com>
RB: 21899