Bug#29739693 CRASH IN ITEM_FIRST_LAST_VALUE::CLEAR()
`reset_all_wf_state' at start of [Buffering]WindowingIterator::Init is
called by JOIN::exec if there are no windowing steps, i.e. the
optimizer has determined that the result set has only a single row: in
this case, the Item_sum_xxx::val_yyy will get the correct value
without any real windowing, as longs as they are properly initialized.
When we have windowing iterators (i.e. we have windowing steps) there
is no need to call this function.

Indeed, it is not always (as in this issue) safe to call
`reset_all_wf_state' since Item_first_last_value::clear assumes the wf
has been prepared.  This is not the case in this issue: the window
itself hasn't been optimized away, but the set of actual windowing
function using that window is zero,
cf. Window::m_outtable_param->items_to_copy, which is empty here.  It
is zero because the window function is only called in an ORDER BY
clause, which itself is optimized away due to the subquery being a
scalar subquery. This meant that the window function didn't get
prepared in the EXECUTE phase (SELECT_LEX::order_list is empty in
EXECUTE's prepare), leading to the crash.

Instead of calling reset_all_wf_state in [Buffering]WindowingIterator,
we need to call call Window::reset_round, which replaces the earlier
call to reset_round in the old execution machinery from JOIN::reset.
This initializes the window for (a new[1]) execution.

[1] in subqueries with windowing which are executed several times, it
is done prior to each execution.

Change-Id: I476f8b5f2acd5eb56e7155787e049b8d8e3c11b1