Bug#29812613 REAL_FREE_SHARE FAULTY CHECK FOR NOT OPEN LIST
The function NDB_SHARE::real_free_share() has a too strict assert
which check that a NDB_SHARE with same key does not exist in the
list of open shares. This is not true as a new NDB_SHARE instance with
same key might have been created already to serve other queries
working with the new or altered table.

There are also two places in NDB_SHARE::mark_share_dropped() which uses
unordered_map::erase() to check that NDB_SHARE does does not exist in
the open and dropped lists respectively. It's not correct to use the
erase() function to check for existence since such a call will have
sideeffects on contents in the list.

Fix by:
1) rewriting the affected functions to more strictly check
   preconditions.
2) remove the check if same key exists in open list while
   releasing the old NDB_SHARE
3) Don't use erase() when checking for existence of key in list

Also remove remnants of old "moving away for safety" supression from
ndb_alter_table_dml.test, such error does not exist in the code anymore.

Change-Id: I4022a32a626347c5392fb7b43ef2d6f50f044626