Bug #29264764 MYSQL ROUTER DEPENDS ON ORDER WITHOUT ORDER BY
The router keeps the list of the cluster nodes that it queries from the metadata
in the vector in the exact order it got from the sql query.
That query does not contain ORDER BY clause so this order is not deterministic
and we rely on the lists queried from different nodes to have the same order
as we compare the instances one-by-one.

This patch changes the used container from vector to set to make sure the
instances are stored always in the same order.

RB: 22163