Bug#11748692: LOGGING TO SLOW_LOG TABLE FAILS WITH VERY SLOW QUERIES
In scenarios where slow-logging to a table overflowed a given field
(e.g. more rows were examined than fit into a 4-byte integer), this
issue was detected, and the writing of the row aborted.

Now we write whatever information we can provide instead (as long
as the table can be opened and used, and the row written; otherwise,
we still skip the line and log the issue to the error log).

Approved by: Jens Even <jens.even.blomsoy@oracle.com>