Bug#29201665: ASAN MEMORY LEAK IN LOG_EVENT::WRITE_HEADER() IN MYSQL-5.7
This fix is a backport of Bug#22732184.

The root cause is that executing 'FLUSH LOGS' tries to write a
Rotate_log_event into an uninitialized log_file. This happens if
there is no binlog file before executing 'FLUSH LOGS'.

To fix the problem, do not write a Rotate_log_event into an
uninitialized log_file in any case.

RB:22182