Bug#29833031 crash on after rest plugins are unloaded
Problem
=======

RestApiComponent crashes at destruct if RestApiComponent users
forget to remove a path they've added.

The RestApiComponent would destruct a unique_ptr<> which points
to the plugin's code which is long unloaded. In most cases, this
doesn't crash, but rarely the memory of the unloaded plugins got
reused which leads to the crash.

Change
======

Add a RestApiComponentPath class which manages the add_path() and
remove_path() for each plugin.

RB: 22229