Bug#29756808 assertion failure: trx0trx.*trx_state_eq(((trx)), trx_state_not_started).*
Problem:
The discrepancy between the data-dictionary and InnoDB storage
engine metadata has occurred In cases when ALTER TABLESPACE ...
ENCRYPTION statement was killed after the operation was started
in SE, but before statement commit. As result server might have
failed on assertion when trying to access this tablespace later.

This problem was especially visible in ntest workload.

The problem is that hton->alter_tablespace() method in InnoDB SE
is implemented in such a way that it expects that the "ALTER
TABLESPACE ... ENCRYPTION" operation never fails after its return.
But this is impossible to achieve since after this call ALTER
TABLESPACE still needs to upgrade MDL lock on tables involved,
update data-dictionary, write statement to the binary log and
commit all these changes. Each of these operations might fail
due to various reasons, e.g. to honour KILL QUERY, due to
running out of disk or memory or simply because server/machine
crashes.

So the proper solution to this problem is to change
hton->alter_tablespace() to support transaction rollback.
However, this change is non-trivial.

Meanwhile, this patch provides workaround for problem with
ntest workload by making ALTER TABLESPACE immune to KILL QUERY
request while we acquire MDL lock on tables in tablespace to
update per table ENCRYPTION clause. We also make this lock
acquisition immune to low timeouts.

The proper solution to this problem is to be implemented in
in the scope of bug@29862604.

Change-Id: Ie79fd0917088d412e8491600c4f38255b5d74d59