Bug#29210560 ADD A REPLICATION TEST FOR CASES COVERED IN NDB_RPL.NDB_RPL_BINLOG _FORMAT_ERRORS
This patch is a backport of fix for Bug#29210560 to mysql-5.6.

Extract the test scenarios for binlog format errors and warnings which do not
depend on NDB storage engine to a new test of rpl_nogtid test suite.

Problem: Test NDB_RPL.NDB_RPL_BINLOG_FORMAT_ERRORS verifies the correctness of
the errors and warning issued for all conditions where the binlog format of a
statement is wrong. Although it covers scenarios which do not depend on NDB
storage engine, it belongs to the ndb_rpl test suite and therefore it is not run
regularly on PB2 mysql server branches.

Solution: Extract the test scenario which do not depend on NDB to a new test of
rpl_nogtid suite. The test NDB_RPL.NDB_RPL_BINLOG_FORMAT_ERRORS is now
responsible for verifying  that a binlog error is issued if more than one engine
is involved in a statement and at least one is doing it's own logging, which is
the case of NDB storage engine.

The new test requires GTIDs to be disabled. The decision on the binlog format
does not depend on the value of GTID_MODE. It is not run in multi-threaded
slave mode since an error in a worker slave leads to an upper-level error in
the coordinator. It cannot run on Windows platform because dynamic loading of
the example plugin does not work.

The scenario where the binlog format is 'statement', the storage engine is
row-only and the default database is ignored by the binlog was commented out due
to a bug. Since the scenario is passing enable it.

During the test cleanup stop the slave before resetting the binlog format since
this operation cannot be done while the slave is running.

RB: 22252
Reviewed-by: Pedro Figueiredo <pedro.figueiredo@oracle.com>
Reviewed-by: Narendra Chauhan <narendra.chauhan@oracle.com>