Bug #29813582 INSUFFICIENT BUFFER SIZE TO HOLD MAX LENGTH NAME OF PARTITIONED TABLE
Problem:

When a partitioned table name is converted to the file name, the buffer used to
hold the file name did not have enough space.  An assert to check whether the
buffer had sufficient space failed.

Solution:

Convert the assert into an if condition and report the error situation to the
caller.

rb#22253 approved by Rahul Agarkar <rahul.agarkar@oracle.com>