Bug #29853165 TCPPORT POOL THAT ROUTER TESTS USE CREATES A DIRECTORY WITH WRONG PERMISSIONS.
TCPPortPool creates a directory for lock-files so that different tests did
not try to use the same ports. It uses the same directory and file scheme as MTR
does. When we create this directory we use 0777 mask which is what it should be.
The problem is that we do not take umask into account which on most systems
will limit the directory rights. That can lead to test failures when someone runs
the tests on a machine and then someone else tries to do that as well and
does not have write rights to the /tmp/mysql-unique-ids

This patch uses chmod to force 0777 after the mkdir call.
Also it offsets the range of the port numbers that Router uses
from 10000-13000 to 11000-13000 to decrease the likelyhood of the clashes
on some commonly used ports (docker 10010 etc.)

RB: 22254