Bug#29825148 NDB TEST: PB2(7.6): NDB_PARALLEL_UNDO_LOG GOT TEMP ERROR 499 SCAN TAKEOVER ERR
Problem: A scan-delete/update of many tuples may fail when the data nodes
are overloaded.

Fix: Extend the test to accomodate other possible errors and limit the
number of retries to end the test when resource problems persist on
the test machines.

Approved by: Sanjana Ds <sanjana.ds@oracle.com>