BUG#29390506 HIDDEN GENERATED COLUMNS PREVENT SLAVE FROM HAVING EXTRA COLUMNS IN TABLES
Description
-----------

After BUG#29317684 when replicating using RBR, it's impossible to replicate data
for tables with extra columns on the slave, when those columns are added after
hidden generated columns - used by functional indexes.

Analysis
--------

BUG#29317684 introduced a change that groups all hidden generated columns at the
end of the columns set - even if the column is added with `ALTER TABLE`.

This prevents the proper replication of tables for which a slave has extra
columns because the order of the columns won't match the one on the master.

```
      MASTER TABLE `t`                SLAVE TABLE `t`
      +----+----+----+------+------+  +----+----+----+-----+------+------+
      | C1 | C2 | C3 | HGC1 | HGC2 |  | C1 | C2 | C3 | EC1 | HGC1 | HGC2 |
      +----+----+----+------+------+  +----+----+----+-----+------+------+
```

In the above example, the extra column `EC1` in the slave will be paired with
the hidden generated column `HGC1` of the master.

Fix
---

Prevent hidden generated tables from being replicated:

1) Implement an iterator interface to abstract the TABLE object field set
   iteration and exclude hidden generated columns from the iteration process.

2) Use the implemented iterator to compose the table metadata in
   `Table_map_log_event`.

3) Use the implemented iterator to pack the row image data in
   `Rows_log_event` and related functions.

4) Use the implemented iterator to unpack the row image data in
   `Rows_log_event` and related functions.

5) Verify the `immediate_server_version` global variable in order to determine
   if the given row should be unpacked with or without filtering.

Reviewed-by: Neha Kumari <neha.n.kumari@oracle.com>
Reviewed-by: Daogang Qu <bill.qu@oracle.com>
RB: 21572