wl#13120 Introduce NdbInterpreter code for comparing two columns from same table.
Patch introduce the new NdbInterpreter instruction BRANCH_ATTR_OP_ATTR, which
compare two Attributes identified as part of the instuction and compare them
with one of the compare operations <, <=, >, >=, <> or =.

Implementation of this instruction build on the similar existing instructions
BRANCH_ATTR_OP_ARG and BRANCH_ATTR_OP_ARG_2 which compare an Attribute against
either a value supplied as part of the BRANCH_ATTR_OP_ARG, or a parameter identified
as part of BRANCH_ATTR_OP_ARG_2.

Variants of the existing NdbScanFilter::cmp() and NdbInterpretedCode::branch_col*()
methods are introduced for generating the NdbInterpreterCode for comparing
two columns from the same table.

Reviewed by : Martin Skold <martin.skold@oracle.com>