Bug#29685643 ELIMINATE SERVER RE-EVALUATION OF PUSHED CONDITION IN A IS_NOT_NULL_COMP TRIGGER
When (parts of) a condition inside a IS_NOT_NULL_COMP trigger condition
has been pushed down to the storage engine, there is not a need for the
mysqld server to re-evaluate the same condition on the (matching) rows
returned from the SE.

Eliminationg such double evaluation of conditions has proven to be
particular important for like-predicates which require considerable
amount of CPU cycles. For the Q13-query in the TPC-H test suite,
we found that over 20% of the CPU cycles were consumed by the
like-condition (used the perf tool). Overall, we improved the Q13
from 4.15s to 3.95s by not doing such double evaluations.

Patch build on the existing metodology in cond_push_boolean_term()
where we already eliminates pushed AND/OR terms from a condition.
If we were able to eliminate all AND/OR terms, the entire AND/OR
condition could be eliminated.

Patch extends this to the IS_NOT_NULL_COMP trigger:
  - If there is remainder condition after pushing the inner
    condition of the IS_NOT_NULL_COMP trigger, the trigger
    is updated with this condition.
  - If there is no such remainder, the entire trigger cond
    has been pushed, and can be eliminated.

Patch has also beenm tested extensively with the RQG (Random Query Generator)
framework.

Reviewed by : Martin Skold <martin.skold@oracle.com>