Bug#29652464 - ORDER OF ADD AND DROP CHECK CONSTRAINT IN TABLE ALTER STATEMENT IS INCORRECT.
In ALTER statement, if both drop and create operations are specified
then first drop operations are performed and then the create. But
for check constraint incorrect order is used. First create
operations are performed and then the drop.

To fix this issue, changed the order of these operations so that
even check constraint follow the same order as other drop and
create operations in ALTER statement.

Reviewed-by: Dmitry Lenev <Dmitry.Lenev@oracle.com>