Bug#29860378 RESULTS NOT SORTED AS SPECIFIED WHEN QUERY IS A 'PUSHED JOIN'
Regression introduced by WL#10234:

The optimizer may decide to use an ordered index to provide the ordering
specified by the ORDER BY clause in a query. If that table is the root
of a pushed join, the result rows may not be returned in the specified
sorted order.

The problem for such SPJ queries are caused by the result effectively
being produced as a cross product of the different (scan) batches
returned from each table in the pushed join. If multiple batches are required
to return all child rows matching a set of parent rows, the rows in the parent batch
are reiterated over the new child batch rows to generate more result rows.
Thus, result rows end up being unsorted, even of the rows in the parent
batch originally were sorted as specified. The original solution to
this (pre wl#10234) was to enforce a batch size of a single row in the
result batch required to be sorted. (At the cost of some performance
compared to larger batches)

WL#10234 introduced a v2 version of the request and parameter format
being sent from the API to the SPJ block: The v2 parameters were extended to
also include a batch size argument, deprecating the batch size previously
set directly in the ScanFragReq structure in 'v1'. Unfortunately the logic
of setting 'batch size = 1' for sorted result was lost in this new v2 logic.

This patch introduce a SORTED_RESULT flag as part of the v2 parameters,
and let the SPJ code set 'batch size = 1' instead of enforcing this on the
API side. Technically the solution is the same as pre-v2, but leaving
the enforcing of 'batch size = 1' to SPJ should be somewhat more future
proof as it would be possible to implement a better (faster) solution
in the SPJ block in the future without changing the API protocol.

Reviewed by : Maitrayi Sabaratnam <maitrayi.sabaratnam@oracle.com>