Bug#27627586 RPL.RPL_ATOMIC_USER FAILS SPORADICALLY ON PB2
In test rpl.rpl_atomic_user exclude from the analysis of the content of
mysql.user table the rows which were created during server startup, to prevent
mismatches between the master and the slave servers due to different startup
times.

Problem: Test RPL.RPL_ATOMIC_USER fails sporadically on PB2. The command
"diff_files" fails with error.

Analysis: The test rpl.rpl_atomic_user executes statements related to account
management, creating users and roles which are added to table mysql.user. In two
of its steps it is verified if the mysql.user table has the same values in both
the master and slave server, by invoking the command diff_tables. This command
selects all the rows of the table for comparison including those users created
during the initialization of the server : root, mysql.session, mysql.sys and
mysql.infoschema. An error can occur if there is a difference in the
initialization between the master and the slave server. In particular if there
is a difference in the startup time greater than the resolution of the field
"password_last_changed" (timestamp indicating when the password was changed) the
test will fail.

Solution: When comparing the content of table mysql.user between the master and
server, exclude the rows of the users created during server initialization.
Since it is currently not possible to use the command diff_tables for tables
which are supposed to be different, create a temporary table which is a copy of
the non-excluded rows of mysql.user and pass it as the argument to diff_tables.

RB: 22294
Reviewed-by: Pedro Figueiredo <pedro.figueiredo@oracle.com>
Reviewed-by: Narendra Chauhan <narendra.chauhan@oracle.com>