Bug#29705858 RUN_NDBAPITEST.INC ALWAYS EXIT TEST CASE WHEN THE NDBAPITEST SUCCEED
Exec_print_on_error() is used to run any executable in the system and upon
success run_ndbapitest.inc forcefully exits the test cases. This prevents the
execution of MTR framework function check_testcase(), which is run at the end
of every test case to check that the state of the database is same before
and after running the test.

Fix by removing the exit() function when exec_print_on_error() is success.

Change-Id: I99e2f57827447da0e4836f1bc197e97f41d1e456