Bug #29876948 MYSQL ROUTER FAILS TO CONNECT TO CLUSTER
The router was using TCPAddress class to crete an URI that was then stored in the state file.
That does not work well if the address happened to be classified as IPv6 which is part of the magic
that TCPAddress class does. If that was the case it put [] around the string representation.

This patch replaces use of TCPAddress with URI object to create an uri
that is stored in the state file.

RB: 22306