Bug#29754242: GR_LARGE_VIEW_CHANGE_EVENT_SIZE SSL FAILS IN PARALLEL TESTING
The test load was too big in some cases causing timeouts.
We cut the test load in half and increase the wait times to prevent
further failures.

The test was also corrected to use slave_max_allowed_packet as the a
small max_allowed_packet value does not cause failures even if the
VCLE size exceeds it.

ReviewBoard: 22311